# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LeadChain API",
        "description": "API REST de LeadChain para la gestión comercial de clientes, edificios y visitas. Autenticación mediante JWT (Bearer token). Incluye soporte para coordenadas geográficas PostGIS.",
        "contact": {
            "name": "Alberto Romero Pino",
            "email": "albertoromeropino2004@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos los clientes",
                "operationId": "de55d50d1b708d9e5c63ee3f8fe8f54a",
                "responses": {
                    "200": {
                        "description": "Listado de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClienteResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un nuevo cliente",
                "operationId": "4ac6cdfa8d6c188fa2cb4f3263d4318a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Antonio"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Perez Garcia"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "antonio@ejemplo.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "612345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/edificios": {
            "get": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Listar todos los edificios",
                "operationId": "99106a5021f84eaf8e2b8107619b7bcd",
                "responses": {
                    "200": {
                        "description": "Listado de edificios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EdificioResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Crear un nuevo edificio",
                "operationId": "3857737f53be84d355ff0ba0754fb283",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Edificio creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EdificioResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al crear el edificio"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/edificios/{edificio}/detalle": {
            "get": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Obtener detalles completos de un edificio",
                "description": "Retorna todo lo necesario para la vista de detalles en una sola consulta: edificio, zona, otros edificios del bloque y todas las zonas",
                "operationId": "ae87bd0a7d12487dd5677e43c3665665",
                "parameters": [
                    {
                        "name": "edificio",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del edificio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EdificioDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Edificio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/edificios/{edificio}": {
            "put": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Actualizar un edificio completo",
                "operationId": "c58986ead2caaa128f644228ea060c97",
                "parameters": [
                    {
                        "name": "edificio",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Edificio actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EdificioResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Eliminar un edificio",
                "operationId": "79d02b8139ffc8389c4d50284c55b021",
                "parameters": [
                    {
                        "name": "edificio",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Edificio eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Actualizar parcialmente un edificio",
                "operationId": "564907a9d9787929b78c5c47226e0fbe",
                "parameters": [
                    {
                        "name": "edificio",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Edificio actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EdificioResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "GeoPoint": {
                "required": [
                    "lat",
                    "lng"
                ],
                "properties": {
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "example": 37.9
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "example": -4.8
                    }
                },
                "type": "object"
            },
...
            "ClienteResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Antonio"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Perez Garcia"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "612345678"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "antonio@ejemplo.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EdificioResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "direccion_completa": {
                        "type": "string",
                        "example": "Calle Gran Capitan 10"
                    },
                    "ubicacion": {
                        "$ref": "#/components/schemas/GeoPoint"
                    },
                    "id_zona": {
                        "type": "integer",
                        "example": 1
                    },
                    "tipo": {
                        "type": "string",
                        "example": "residencial"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ZonaResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Zona Norte"
                    },
                    "area": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GeoPoint"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "EdificioDetailResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "direccion_completa": {
                        "type": "string",
                        "example": "Calle Gran Capitan 10"
                    },
                    "ubicacion": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/GeoPoint"
                            }
                        ],
                        "nullable": true
                    },
                    "id_zona": {
                        "type": "integer",
                        "example": 1
                    },
                    "tipo": {
                        "type": "string",
                        "example": "residencial"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "clientes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ClienteResource"
                        }
                    },
                    "zona": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/ZonaResource"
                            }
                        ],
                        "nullable": true
                    },
                    "bloqueEdificios": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EdificioDetailResource"
                        }
                    },
                    "todasLasZonas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ZonaResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}